Cleaned up TR/GGX code a bit
